Update README.md
It also provides information on formatting (how to make a header, italic and bold font, and bulleted lists), how to make tables, dealing with dates, converting characters to numbers, renaming or removing columns, and adding month names to the horizontal axis of figures.  

Some of these actions are not intuitive!